Fix error causing newline chars not to print
Pattern matched function heads in elixir are tested in order
so having:
  def fits(w,%Text{text: s, doc: x}), do: fits((w - String.length s),x)
  def fits(_,%Line{indent: _,doc: _}), do: true
  def fits(_,:Nil), do: true
  def fits(w,_) when w < 0, do: false
Meant that the final fits case would never be called when the available
line space was already exceded. This commit places that final
pattern match head at the top of the list so if line space is exceded
fits will return false regardless of the contents of the rest of the
line.